LayoutTests/crypto/subtle/derive-key-for-all-supported-algorithms.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script type="text/javascript">
description("We test deriveKey and then make sure that ecdh produces the same shared secret and the same encryption results using a key derived from that secret.");

/**
 * Copy and paste the below code to generate `keyData` for the test.

```bash
node << EOF
function bytesToHexString(bytes)
{
    if (!bytes)
        return null;

    bytes = new Uint8Array(bytes);
    var hexBytes = [];

    for (var i = 0; i < bytes.length; ++i) {
        var byteString = bytes[i].toString(16);
        if (byteString.length < 2)
            byteString = "0" + byteString;
        hexBytes.push(byteString);
    }

    return hexBytes.join("");
}

async function generateKeys() {
    let keyPair1 = await crypto.subtle.generateKey({'name': "ECDSA", 'namedCurve':'P-256'}, true, ['sign', 'verify']);
    let keyPair2 = await crypto.subtle.generateKey({'name': "ECDSA", 'namedCurve':'P-256'}, true, ['sign', 'verify']);
    let private1 = await crypto.subtle.exportKey("pkcs8", keyPair1.privateKey);
    let public1 = await crypto.subtle.exportKey("raw", keyPair1.publicKey);
    return [
        bytesToHexString(await crypto.subtle.exportKey("pkcs8", keyPair1.privateKey)),
        bytesToHexString(await crypto.subtle.exportKey("raw", keyPair1.publicKey)),
        bytesToHexString(await crypto.subtle.exportKey("pkcs8", keyPair2.privateKey)),
        bytesToHexString(await crypto.subtle.exportKey("raw", keyPair2.publicKey))
    ];
}

let keys = await generateKeys();
keys.forEach((key) => console.log('   hexStringToUint8Array("' + key + '"),'));
EOF

```

*/

jsTestIsAsync = true;
let keyData = [
    hexStringToUint8Array("308187020100301306072a8648ce3d020106082a8648ce3d030107046d306b0201010420fe77a808a7109ba5ceb93ebebad2c84a714d864ad29b62d6537e1969035c0079a144034200042684c752eef1c927a80c74e8b02ce459f848b5977f37fd878b36dae632be9a6cadd56126e404a4f75c535e5769d95b49fb1106f784f3d231b776d1f4d57927ce"),
    hexStringToUint8Array("042684c752eef1c927a80c74e8b02ce459f848b5977f37fd878b36dae632be9a6cadd56126e404a4f75c535e5769d95b49fb1106f784f3d231b776d1f4d57927ce"),
    hexStringToUint8Array("308187020100301306072a8648ce3d020106082a8648ce3d030107046d306b020101042067521ccd1f85516118182bca3394c273bab9ce5cd6265105559e325e01f2df1ca144034200043042d8698882f2b59de972390d3fc9277e2e677a6c560148017c9475218fda1b38f76f7645fbcaf3d03e6259d080204fbafb04731b6ad53cb25c3d35d95b7c73"),
    hexStringToUint8Array("043042d8698882f2b59de972390d3fc9277e2e677a6c560148017c9475218fda1b38f76f7645fbcaf3d03e6259d080204fbafb04731b6ad53cb25c3d35d95b7c73"),
];
var allDerivedKeys = [];
var allEncryptionKeys = [];
var allEncryptionResults = [];
let iv = new Uint8Array(Array(12).keys());
let salt = new Uint8Array(Array(10).keys());
let plaintext = new Uint8Array(Array(100).keys());
async function runTest() {
    let extractable = true;
    var allKeys = await Promise.all([
        crypto.subtle.importKey("pkcs8", keyData[0], {name: "ECDH", namedCurve: "P-256"}, extractable, ["deriveKey", 'deriveBits']),
        crypto.subtle.importKey("raw", keyData[1], {name: "ECDH", namedCurve: "P-256"}, extractable, []),
        crypto.subtle.importKey("pkcs8", keyData[2], {name: "ECDH", namedCurve: "P-256"}, extractable, ["deriveKey", 'deriveBits']),
        crypto.subtle.importKey("raw", keyData[3], {name: "ECDH", namedCurve: "P-256"}, extractable, []),
    ]); 
    allDerivedKeys = await Promise.all([
        crypto.subtle.deriveKey({name: "ECDH", public: allKeys[3] }, allKeys[0], { name: "HKDF", hash: "" , salt: new Uint8Array(), info: new Uint8Array() }, false, ["deriveKey"]),
        crypto.subtle.deriveKey({name: "ECDH", public: allKeys[3] }, allKeys[0], { name: "PBKDF2", hash: "" , salt: new Uint8Array(), iterations: 32 }, false, ["deriveKey"]),
        crypto.subtle.deriveKey({name: "ECDH", public: allKeys[1] }, allKeys[2], { name: "HKDF", hash: "" , salt: new Uint8Array(), info: new Uint8Array() }, false, ["deriveKey"]),
        crypto.subtle.deriveKey({name: "ECDH", public: allKeys[1] }, allKeys[2], { name: "PBKDF2", hash: "" , salt: new Uint8Array(), iterations: 32 }, false, ["deriveKey"]),
    ]);
    console.log("ecdh keys done");
    allEncryptionKeys = await Promise.all([
        crypto.subtle.deriveKey({name: "HKDF", hash: "SHA-256", salt: salt, info: plaintext}, allDerivedKeys[0], {name:"AES-GCM", length: 256}, true, ["encrypt", "decrypt"]),
        crypto.subtle.deriveKey({name: "PBKDF2", hash: "SHA-256", salt: salt, iterations: 32 }, allDerivedKeys[1], { name:"AES-GCM", length: 256 }, true, ["encrypt", "decrypt"]),
        crypto.subtle.deriveKey({name: "HKDF", hash: "SHA-256", salt: salt, info: plaintext}, allDerivedKeys[2], {name:"AES-GCM", length: 256}, true, ["encrypt", "decrypt"]),
        crypto.subtle.deriveKey({name: "PBKDF2", hash: "SHA-256", salt: salt, iterations: 32 }, allDerivedKeys[3], { name:"AES-GCM", length: 256 }, true, ["encrypt", "decrypt"]),
    ]);
    console.log("aes keys done");
    allEncryptionResults = await Promise.all([
        crypto.subtle.encrypt({ name: "AES-GCM", iv: iv }, allEncryptionKeys[0], plaintext),
        crypto.subtle.encrypt({ name: "AES-GCM", iv: iv }, allEncryptionKeys[1], plaintext),
        crypto.subtle.encrypt({ name: "AES-GCM", iv: iv }, allEncryptionKeys[2], plaintext),
        crypto.subtle.encrypt({ name: "AES-GCM", iv: iv }, allEncryptionKeys[3], plaintext)
    ]);
    console.log("all done");
}
runTest().then(() => {
    shouldBe('bytesToHexString(allEncryptionResults[0])', '"a6280c522670eaf82f6564afbeb20a5b3f2d4e13c5596f6df3dcff8c34cb2118d2770fb24d83cfac5079c323118485bb01170292ee41eb82b07208f4840478fea3771d8922785c476ba06c2a0b933fc1661431419530a916ad4468545d1af5004a1149fea241c2ff1582ee58a8b7d79935de5def"');
    shouldBe('bytesToHexString(allEncryptionResults[1])', '"c6201dfbb6fa92c1c246f6ce52f8f1c037f087efde41bac7f6485a2a8207623d2d3825b9cbe8ef864a90378667ed25544ce44cd2904bd96c19f0eeb611d626185165a8afb4e52f95700d7880f83939a42712fc4e377f198c01a61b397b76c3a4b93d932c321084bbef33332169dea09458b27df3"');
    shouldBe('bytesToHexString(allEncryptionResults[0])', 'bytesToHexString(allEncryptionResults[2])');
    shouldBe('bytesToHexString(allEncryptionResults[1])', 'bytesToHexString(allEncryptionResults[3])');
    console.log(bytesToHexString(allEncryptionResults[0]));
    console.log(bytesToHexString(allEncryptionResults[1]));
    console.log(bytesToHexString(allEncryptionResults[2]));
    console.log(bytesToHexString(allEncryptionResults[3]));
    finishJSTest();
}, (e) => { console.log(e); finishJSTest()});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
