LayoutTests/crypto/subtle/derive-key-for-all-supported-algorithms-expected.txt

CONSOLE MESSAGE: ecdh keys done
CONSOLE MESSAGE: aes keys done
CONSOLE MESSAGE: all done
CONSOLE MESSAGE: a6280c522670eaf82f6564afbeb20a5b3f2d4e13c5596f6df3dcff8c34cb2118d2770fb24d83cfac5079c323118485bb01170292ee41eb82b07208f4840478fea3771d8922785c476ba06c2a0b933fc1661431419530a916ad4468545d1af5004a1149fea241c2ff1582ee58a8b7d79935de5def
CONSOLE MESSAGE: c6201dfbb6fa92c1c246f6ce52f8f1c037f087efde41bac7f6485a2a8207623d2d3825b9cbe8ef864a90378667ed25544ce44cd2904bd96c19f0eeb611d626185165a8afb4e52f95700d7880f83939a42712fc4e377f198c01a61b397b76c3a4b93d932c321084bbef33332169dea09458b27df3
CONSOLE MESSAGE: a6280c522670eaf82f6564afbeb20a5b3f2d4e13c5596f6df3dcff8c34cb2118d2770fb24d83cfac5079c323118485bb01170292ee41eb82b07208f4840478fea3771d8922785c476ba06c2a0b933fc1661431419530a916ad4468545d1af5004a1149fea241c2ff1582ee58a8b7d79935de5def
CONSOLE MESSAGE: c6201dfbb6fa92c1c246f6ce52f8f1c037f087efde41bac7f6485a2a8207623d2d3825b9cbe8ef864a90378667ed25544ce44cd2904bd96c19f0eeb611d626185165a8afb4e52f95700d7880f83939a42712fc4e377f198c01a61b397b76c3a4b93d932c321084bbef33332169dea09458b27df3
We test deriveKey and then make sure that ecdh produces the same shared secret and the same encryption results using a key derived from that secret.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(allEncryptionResults[0]) is "a6280c522670eaf82f6564afbeb20a5b3f2d4e13c5596f6df3dcff8c34cb2118d2770fb24d83cfac5079c323118485bb01170292ee41eb82b07208f4840478fea3771d8922785c476ba06c2a0b933fc1661431419530a916ad4468545d1af5004a1149fea241c2ff1582ee58a8b7d79935de5def"
PASS bytesToHexString(allEncryptionResults[1]) is "c6201dfbb6fa92c1c246f6ce52f8f1c037f087efde41bac7f6485a2a8207623d2d3825b9cbe8ef864a90378667ed25544ce44cd2904bd96c19f0eeb611d626185165a8afb4e52f95700d7880f83939a42712fc4e377f198c01a61b397b76c3a4b93d932c321084bbef33332169dea09458b27df3"
PASS bytesToHexString(allEncryptionResults[0]) is bytesToHexString(allEncryptionResults[2])
PASS bytesToHexString(allEncryptionResults[1]) is bytesToHexString(allEncryptionResults[3])
PASS successfullyParsed is true

TEST COMPLETE
